mv methods to handler.go
